Fix storybook cannot show translated message
